Move the languge at home editing to the additional info section
