+++ frontend/src/employee-frontend/components/person-shared/PersonDetails.tsx
@@ -2,7 +2,6 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import sortBy from 'lodash/sortBy'
 import React, {
   useCallback,
   useContext,
@@ -17,15 +16,13 @@ import { Loading, Result, Success } from 'lib-common/api'
 import { UpdateStateFn } from 'lib-common/form-state'
 import { Action } from 'lib-common/generated/action'
 import { PersonJSON } from 'lib-common/generated/api-types/pis'
-import { IsoLanguage, isoLanguages } from 'lib-common/generated/language'
+import { isoLanguages } from 'lib-common/generated/language'
 import LocalDate from 'lib-common/local-date'
 import Button from 'lib-components/atoms/buttons/Button'
 import InlineButton from 'lib-components/atoms/buttons/InlineButton'
-import Combobox from 'lib-components/atoms/dropdowns/Combobox'
 import Checkbox from 'lib-components/atoms/form/Checkbox'
 import InputField from 'lib-components/atoms/form/InputField'
 import Radio from 'lib-components/atoms/form/Radio'
-import TextArea from 'lib-components/atoms/form/TextArea'
 import { SpinnerSegment } from 'lib-components/atoms/state/Spinner'
 import {
   FixedSpaceColumn,
@@ -87,8 +84,6 @@ interface Form {
   invoicingPostOffice: string
   forceManualFeeDecisions: boolean
   ophPersonOid: string
-  languageAtHome: string
-  languageAtHomeDetails: string
 }
 
 const RightAlignedRow = styled.div`
@@ -100,14 +95,6 @@ const RightAlignedRow = styled.div`
 const ButtonSpacer = styled.div`
   margin-right: 25px;
 `
-
-const filterLanguages = (
-  input: string,
-  items: IsoLanguage[]
-): IsoLanguage[] => {
-  const filter = input.toLowerCase()
-  return items.filter((item) => item.nameFi.includes(filter))
-}
 
 export default React.memo(function PersonDetails({
   person,
@@ -134,9 +121,7 @@ export default React.memo(function PersonDetails({
     invoicingPostalCode: '',
     invoicingPostOffice: '',
     forceManualFeeDecisions: false,
-    ophPersonOid: '',
-    languageAtHome: '',
-    languageAtHomeDetails: ''
+    ophPersonOid: ''
   })
   const [ssnDisableRequest, setSsnDisableRequest] = useState<Result<void>>(
     Success.of()
@@ -184,9 +169,7 @@ export default React.memo(function PersonDetails({
         invoicingPostalCode: person.invoicingPostalCode ?? '',
         invoicingPostOffice: person.invoicingPostOffice ?? '',
         forceManualFeeDecisions: person.forceManualFeeDecisions ?? false,
-        ophPersonOid: person.ophPersonOid ?? '',
-        languageAtHome: person.languageAtHome,
-        languageAtHomeDetails: person.languageAtHomeDetails
+        ophPersonOid: person.ophPersonOid ?? ''
       })
     }
   }, [person, editing])
@@ -246,18 +229,6 @@ export default React.memo(function PersonDetails({
         : null
       )?.nameFi ?? person.language,
     [person.language]
-  )
-
-  const languageAtHome = useMemo(
-    () =>
-      (person.languageAtHome ? isoLanguages[person.languageAtHome] : null) ??
-      null,
-    [person.languageAtHome]
-  )
-
-  const languages = useMemo(
-    () => sortBy(Object.values(isoLanguages), ({ nameFi }) => nameFi),
-    []
   )
 
   return (
@@ -503,53 +474,6 @@ export default React.memo(function PersonDetails({
                 }
               ]
             : []),
-          {
-            label: i18n.childInformation.personDetails.languageAtHome,
-            value: editing ? (
-              <>
-                <Combobox
-                  data-qa="input-language-at-home"
-                  items={languages}
-                  getItemDataQa={(item) => `language-${item.id}`}
-                  selectedItem={
-                    (form.languageAtHome
-                      ? isoLanguages[form.languageAtHome]
-                      : null) ?? null
-                  }
-                  onChange={(item) =>
-                    updateForm({ languageAtHome: item?.id ?? '' })
-                  }
-                  filterItems={filterLanguages}
-                  getItemLabel={(item) => item?.nameFi}
-                  placeholder={
-                    i18n.childInformation.personDetails.placeholder
-                      .languageAtHome
-                  }
-                  clearable={true}
-                />
-                <TextArea
-                  data-qa="input-language-at-home-details"
-                  value={form.languageAtHomeDetails}
-                  placeholder={
-                    i18n.childInformation.personDetails.placeholder
-                      .languageAtHomeDetails
-                  }
-                  onChange={(languageAtHomeDetails) =>
-                    updateForm({ languageAtHomeDetails })
-                  }
-                />
-              </>
-            ) : (
-              <>
-                <div data-qa="person-language-at-home">
-                  {languageAtHome?.nameFi ?? ''}
-                </div>
-                <div data-qa="person-language-at-home-details">
-                  {person.languageAtHomeDetails}
-                </div>
-              </>
-            )
-          },
           ...(!isChild && permittedActions.has('READ_INVOICE_ADDRESS')
             ? [
                 {
